[feat] MAJOR: First MVP of non-XState
