--- src/machine/index.ts
@@ -1,3 +1,0 @@
-import machine from './machine';
-
-export default machine;
